Finished user profile page, refined other pages

+++ src/pages/UserProfile.vue
@@ -10,19 +10,29 @@
         <el-divider></el-divider>
         <h3>My Articles</h3>
         <el-table :data="articles" style="width: 100%;">
-            <el-table-column prop="title" min-width="300" label="Title"></el-table-column>
-            <el-table-column prop="created_at" min-width="100" label="Create Time"></el-table-column>
-            <el-table-column prop="updated_at" min-width="100" label="Update Time"></el-table-column>
+            <el-table-column prop="title" min-width="250" label="Title"></el-table-column>
+            <el-table-column prop="created_at" min-width="130" label="Create Time"></el-table-column>
+            <el-table-column prop="updated_at" min-width="130" label="Update Time"></el-table-column>
             <el-table-column
                     fixed="right"
                     label="Operation"
-                    width="100">
+                    width="150">
                 <template slot-scope="scope">
+                    <el-button @click="clickView(scope.row.id)" type="text" size="small">View</el-button>
                     <el-button @click="clickEdit(scope.row.id)" type="text" size="small">Edit</el-button>
                     <el-button @click="clickDelete(scope.row.id)" type="text" size="small">Delete</el-button>
                 </template>
             </el-table-column>
         </el-table>
+        <el-pagination
+                layout="prev, pager, next"
+                :current-page="this.page"
+                @current-change="this.onPageChange"
+                :page-size="this.pageSize"
+                :total="this.total"
+                style="margin-top: 20px">
+        </el-pagination>
+        <el-button type="success" plain @click="this.clickNewArticle" style="margin-top: 18px">New Article</el-button>
     </el-card>
 </template>
 <script>
@@ -37,17 +47,20 @@
                     email: "email@columbia.edu",
                 },
                 articles: [{
-                    id: 1,
+                    id: 2,
                     title: "test-1",
                     created_at: 1,
                     updated_at: 2
-                }]
+                }],
+                page: 1,
+                pageSize: 5,
+                total: 0
             }
         },
         methods: {
             getUserProfile() {
                 let userId = this.$store.state.userId
-                axios.get(configJson.endpoint + '/api/v1/user/' + userId)
+                axios.get(configJson.endpoint + '/api/v1/users/' + userId)
                     .then(this.getUserProfileSuccess)
                     .catch(function (err) {
                         console.log(err)
@@ -55,34 +68,65 @@
             },
             getUserProfileSuccess(res) {
                 console.log(res)
-                this.profile = res.data
-                this.listArticlesByUserId(res.data.id)
+                this.profile = res.data.user
+                this.listArticlesOfUser()
             },
-            listArticlesByUserId(userId) {
-                // TODO: implement this method
-
-            },
-            listArticlesByUserIdSuccess(res) {
-                this.articles = res.data
-            },
-            clickEdit(row) {
-                console.log(row)
-                console.log(row.id)
-                // TODO: implement this method
-            },
-            clickDelete(articleId) {
-                // TODO: implement this method
-                axios.delete(configJson.endpoint + '/api/v1/articles/' + articleId)
-                    .then(this.deleteArticleSuccess)
+            listArticlesOfUser() {
+                let userId = this.$store.state.userId
+                let url = configJson.endpoint + '/api/v1/articles?page=' + (this.page - 1) + '&limit=' + this.pageSize
+                    + '&author_id=' + userId
+                axios.get(url)
+                    .then(this.listArticlesOfUserSuccess)
                     .catch(function (err) {
                         console.log(err)
                     })
             },
+            listArticlesOfUserSuccess(res) {
+                console.log(res.data)
+                this.articles = res.data.articles
+                this.total = res.data.total
+            },
+            clickView(articleId) {
+                this.$router.push('/articles/' + articleId)
+            },
+            clickEdit(articleId) {
+                console.log('clicked on edit button for article ' + articleId)
+                this.$router.push('/articles/edit/' + articleId)
+            },
+            clickDelete(articleId) {
+                this.$confirm('Are you sure to delete this article?', 'Warning', {
+                    confirmButtonText: 'Yes',
+                    cancelButtonText: 'No',
+                    type: 'warning'
+                }).then(() => {
+                    axios.delete(configJson.endpoint + '/api/v1/articles/' + articleId)
+                        .then(this.deleteArticleSuccess)
+                        .catch(function (err) {
+                            console.log(err)
+                        })
+                }).catch(() => {
+                    this.$notify({
+                        type: 'info',
+                        message: 'Discarded'
+                    });
+                });
+
+            },
             deleteArticleSuccess(res) {
                 console.log(res)
+                this.$notify({
+                    type: 'success',
+                    message: 'Successfully deleted'
+                });
                 // refresh the articles list
-                let userId = this.$store.state.userId
-                this.listArticlesByUserId(userId)
+                this.listArticlesOfUser()
+            },
+            clickNewArticle() {
+                this.$router.push('/articles/edit')
+            },
+            onPageChange(page) {
+                this.page = page
+                this.listArticlesOfSection(this.activeSectionId)
             }
         },
         mounted: function () {
